refactor(validation): Validator.js를 validation.js로 변경 후 utils 디렉토리로 이동
함수형 프로그래밍으로 구현하여 utils에서 관리하는 것이 적합하다고 판단
